Safari-611.1.10/LayoutTests/http/tests/appcache/local-content-expected.txt

CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: Not allowed to load local resource: stdio.h
Test that documents loaded from application cache don't get access to local resources accidentally.

Should say SUCCESS:

SUCCESS
